+++ src/tools/enumErr.ts
@@ -1,4 +1,8 @@
-export function enumLog(params: never): null {
+import { OutputChannel } from '../provider/vscWindows/OutputChannel';
+
+export function enumLog(params: never): void {
     console.log('enumLog ~ params', params);
-    return null;
+    OutputChannel.appendLine('enumLog ~ params');
+    OutputChannel.show();
+    throw new Error('enumLog');
 }